+++ test/payments/p2pkh_test.dart
@@ -1,89 +1,58 @@
-import 'package:coinslib/src/payments/index.dart' show PaymentData;
 import 'package:coinslib/src/payments/p2pkh.dart';
 import 'package:test/test.dart';
-import 'package:coinslib/src/utils/script.dart' as bscript;
-import 'dart:io';
-import 'dart:convert';
 import 'package:hex/hex.dart';
 import 'dart:typed_data';
+import 'package:coinslib/src/models/networks.dart';
 
 main() {
-  final fixtures = json.decode(
-    File("./test/fixtures/p2pkh.json").readAsStringSync(encoding: utf8),
-  );
-
-  group('(valid case)', () {
-    for (final f in (fixtures["valid"] as List<dynamic>)) {
-      test(f['description'] + ' as expected', () {
-        final arguments = _preformPaymentData(f['arguments']);
-        final p2pkh = P2PKH(data: arguments);
-        if (arguments.address == null) {
-          expect(p2pkh.data.address, f['expected']['address']);
-        }
-        if (arguments.hash == null) {
-          expect(_toString(p2pkh.data.hash), f['expected']['hash']);
-        }
-        if (arguments.pubkey == null) {
-          expect(_toString(p2pkh.data.pubkey), f['expected']['pubkey']);
-        }
-        if (arguments.input == null) {
-          expect(_toString(p2pkh.data.input), f['expected']['input']);
-        }
-        if (arguments.output == null) {
-          expect(_toString(p2pkh.data.output), f['expected']['output']);
-        }
-        if (arguments.signature == null) {
-          expect(_toString(p2pkh.data.signature), f['expected']['signature']);
-        }
-      });
-    }
+  test("P2PKH.fromPublicKeyHash", () {
+    final p2pkh = P2PKH.fromPublicKeyHash(
+      HEX.decode("168b992bcfc44050310b3a94bd0771136d0b28d1") as Uint8List,
+    );
+    expect(
+      p2pkh.address(bitcoin),
+      "134D6gYy8DsR5m4416BnmgASuMBqKvogQh",
+    );
+    expect(
+      p2pkh.outputScript,
+      HEX.decode("76a914168b992bcfc44050310b3a94bd0771136d0b28d188ac")
+          as Uint8List,
+    );
   });
-  group('(invalid case)', () {
-    for (final f in (fixtures["invalid"] as List<dynamic>)) {
-      test(
-          'throws ${f['exception']}${f['description'] != null ? ('for ${f['description']}') : ''}',
-          () {
-        final arguments = _preformPaymentData(f['arguments']);
-        try {
-          expect(P2PKH(data: arguments), isArgumentError);
-        } catch (err) {
-          expect((err as ArgumentError).message, f['exception']);
-        }
-      });
-    }
+
+  test("P2PKH.fromPublicKey", () {
+    final p2pkh = P2PKH.fromPublicKey(
+      HEX.decode(
+        "030000000000000000000000000000000000000000000000000000000000000001",
+      ) as Uint8List,
+    );
+    expect(
+      p2pkh.address(bitcoin),
+      "134D6gYy8DsR5m4416BnmgASuMBqKvogQh",
+    );
+    expect(
+      p2pkh.pubKeyHash,
+      HEX.decode("168b992bcfc44050310b3a94bd0771136d0b28d1") as Uint8List,
+    );
+    expect(
+      p2pkh.outputScript,
+      HEX.decode("76a914168b992bcfc44050310b3a94bd0771136d0b28d188ac")
+          as Uint8List,
+    );
   });
-}
 
-PaymentData _preformPaymentData(dynamic x) {
-  final address = x['address'];
-  final hash = x['hash'] != null ? HEX.decode(x['hash']) : null;
-  final input = x['input'] != null ? bscript.fromASM(x['input']) : null;
-  final output = x['output'] != null
-      ? bscript.fromASM(x['output'])
-      : x['outputHex'] != null
-          ? HEX.decode(x['outputHex'])
-          : null;
-  final pubkey = x['pubkey'] != null ? HEX.decode(x['pubkey']) : null;
-  final signature = x['signature'] != null ? HEX.decode(x['signature']) : null;
-  return PaymentData(
-    address: address,
-    hash: hash as Uint8List?,
-    input: input,
-    output: output as Uint8List?,
-    pubkey: pubkey as Uint8List?,
-    signature: signature as Uint8List?,
-  );
-}
-
-String? _toString(dynamic x) {
-  if (x == null) {
-    return null;
-  }
-  if (x is Uint8List) {
-    return HEX.encode(x);
-  }
-  if (x is List<dynamic>) {
-    return bscript.toASM(x);
-  }
-  return '';
+  test("invalid P2PKH hash length", () {
+    expect(
+      () => P2PKH.fromPublicKeyHash(
+        HEX.decode("168b992bcfc44050310b3a94bd0771136d0b28") as Uint8List,
+      ),
+      throwsA(
+        predicate(
+          (e) =>
+              e is ArgumentError &&
+              e.message == 'Invalid P2PKH public key hash length',
+        ),
+      ),
+    );
+  });
 }